--- a/package.json
+++ b/package.json
@@ -17,6 +17,7 @@
     "varint": "^4.0.0"
   },
   "devDependencies": {
+    "fastbench": "^1.0.0",
     "memdb": "^1.0.1",
     "standard": "^5.3.1",
     "tape": "^4.2.1"