--- a/package.json
+++ b/package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "dependencies": {
     "debug": "^2.2.0",
+    "fastparallel": "^2.0.1",
     "framed-hash": "^1.1.0",
     "lexicographic-integer": "^1.1.0",
     "numeric-id-map": "^1.1.0",