As titled :)

Before:

append*10000: 1683ms
writeMemdb*10000: 144ms

After:

append*10000: 1482ms
writeMemdb*10000: 113ms

Possibly there are a ton more of optimizations that can be done, I'm just scratching the surface here (timeboxed 1 hour).