@mcollina 2015-11-25
Here it is now:

append*10000: 1016ms
appendStream*10000: 743ms
writeMemdb*10000: 72ms

the appendStream thing is faster, but still 10x slower. 

Currently it is allocating a bunch of functions over and over again:

```
[marking 0x3e261c50991 <JS Function done (SharedFunctionInfo 0x282069476b39)> for recompilation, reason: hot and stable, ICs with typeinfo: 24/36 (66%), generic ICs: 1/36 (2%)]
[marking 0x3e261cb0b19 <JS Function ready (SharedFunctionInfo 0x282069474e51)> for recompilation, reason: small function, ICs with typeinfo: 3/4 (75%), generic ICs: 0/4 (0%)]
[marking 0x3e261cb1451 <JS Function nodeUpdated (SharedFunctionInfo 0x28206947ac89)> for recompilation, reason: small function, ICs with typeinfo: 5/6 (83%), generic ICs: 0/6 (0%)]
[marking 0x3e261cb13c1 <JS Function insertNode (SharedFunctionInfo 0x282069476a91)> for recompilation, reason: small function, ICs with typeinfo: 10/13 (76%), generic ICs: 0/13 (0%)]
[marking 0x3e261cb0b99 <JS Function gotNodes (SharedFunctionInfo 0x282069476681)> for recompilation, reason: small function, ICs with typeinfo: 4/8 (50%), generic ICs: 1/8 (12%)]
[marking 0x3e261cb07c1 <JS Function (SharedFunctionInfo 0x28206946cd31)> for recompilation, reason: small function, ICs with typeinfo: 4/5 (80%), generic ICs: 0/5 (0%)]
[marking 0x3e261cb1a01 <JS Function (SharedFunctionInfo 0x28206947e2c9)> for recompilation, reason: small function, ICs with typeinfo: 4/6 (66%), generic ICs: 1/6 (16%)]
[marking 0x3e261cb3f59 <JS Function callNext (SharedFunctionInfo 0x282069478c19)> for recompilation, reason: not much type info but very hot, ICs with typeinfo: 0/1 (0%), generic ICs: 1/1 (100%)]
[marking 0x3e261e43559 <JS Function ready (SharedFunctionInfo 0x282069474e51)> for recompilation, reason: small function, ICs with typeinfo: 3/4 (75%), generic ICs: 0/4 (0%)]
[marking 0x3e261e43e91 <JS Function nodeUpdated (SharedFunctionInfo 0x28206947ac89)> for recompilation, reason: small function, ICs with typeinfo: 5/6 (83%), generic ICs: 0/6 (0%)]
[marking 0x3e261e43e01 <JS Function insertNode (SharedFunctionInfo 0x282069476a91)> for recompilation, reason: small function, ICs with typeinfo: 10/13 (76%), generic ICs: 0/13 (0%)]
[marking 0x3e261e435d9 <JS Function gotNodes (SharedFunctionInfo 0x282069476681)> for recompilation, reason: small function, ICs with typeinfo: 4/8 (50%), generic ICs: 1/8 (12%)]
[marking 0x3e261e43201 <JS Function (SharedFunctionInfo 0x28206946cd31)> for recompilation, reason: small function, ICs with typeinfo: 4/5 (80%), generic ICs: 0/5 (0%)]
[marking 0x3e261e44441 <JS Function (SharedFunctionInfo 0x28206947e2c9)> for recompilation, reason: small function, ICs with typeinfo: 4/6 (66%), generic ICs: 1/6 (16%)]
[marking 0x3e261e46a29 <JS Function callNext (SharedFunctionInfo 0x282069478c19)> for recompilation, reason: not much type info but very hot, ICs with typeinfo: 0/1 (0%), generic ICs: 1/1 (100%)]
[marking 0x3e261fd25a1 <JS Function ready (SharedFunctionInfo 0x282069474e51)> for recompilation, reason: small function, ICs with typeinfo: 3/4 (75%), generic ICs: 0/4 (0%)]
[marking 0x3e261fd2ed9 <JS Function nodeUpdated (SharedFunctionInfo 0x28206947ac89)> for recompilation, reason: small function, ICs with typeinfo: 5/6 (83%), generic ICs: 0/6 (0%)]
[marking 0x3e261fd2e49 <JS Function insertNode (SharedFunctionInfo 0x282069476a91)> for recompilation, reason: small function, ICs with typeinfo: 10/13 (76%), generic ICs: 0/13 (0%)]
[marking 0x3e261fd2621 <JS Function gotNodes (SharedFunctionInfo 0x282069476681)> for recompilation, reason: small function, ICs with typeinfo: 4/8 (50%), generic ICs: 1/8 (12%)]
[marking 0x3e261fd2249 <JS Function (SharedFunctionInfo 0x28206946cd31)> for recompilation, reason: small function, ICs with typeinfo: 4/5 (80%), generic ICs: 0/5 (0%)]
[marking 0x3e261fd3489 <JS Function (SharedFunctionInfo 0x28206947e2c9)> for recompilation, reason: small function, ICs with typeinfo: 4/6 (66%), generic ICs: 1/6 (16%)]
[marking 0x3e261fd5951 <JS Function callNext (SharedFunctionInfo 0x282069478c19)> for recompilation, reason: not much type info but very hot, ICs with typeinfo: 0/1 (0%), generic ICs: 1/1 (100%)]
```

The two anonymous functions are from levelup and memdown, so don't worry about those for now :).